requirements.txt
  #Python Modules
  requests>=2.32.5
  python-dotenv>=1.1.1

  #Notion API
  notion-client>=2.5.0

  # Youtube API
  google-api-python-client>=2.185.0
  google-auth-oauthlib>=1.2.2
  google-auth-httplib2>=0.2.0